+++ src/includes/dbquery.php
@@ -33,8 +33,10 @@ class UProfileQuery {
 
     public static function Profile(Ab_Database $db, $userid){
         $sql = "
-			SELECT u.*
+			SELECT 
+			    u.*, p.*
 			FROM ".$db->prefix."user u
+			LEFT JOIN ".$db->prefix."uprofile p ON u.userid=p.userid 
 			WHERE u.userid=".bkint($userid)."
 			LIMIT 1
 		";
@@ -92,98 +94,4 @@ class UProfileQuery {
 		";
         $db->query_write($sql);
     }
-
-    public static function FieldList(Ab_Database $db){
-        $sql = "
-			SELECT
-				fieldid as id, 
-				fieldname as nm,
-				fieldtype as ft,
-				fieldaccess as fa,
-				options as ops,
-				title as tl,
-				ord
-			FROM ".$db->prefix."upfl_field
-			WHERE fieldaccess<>".UserFieldAccess::SYSTEM."
-			ORDER BY ord
-		";
-        return $db->query_read($sql);
-    }
-
-    public static function FieldAppend(Ab_Database $db, $name, $title, $type, $size = '1', $options){
-        $name = bkstr($name);
-        $title = bkstr($title);
-        $size = bkstr($size);
-        $unsignedStr = $options['unsigned'] ? 'unsigned' : '';
-        $defvalue = bkstr($options['default']);
-        $sql = "";
-        switch ($type){
-            case UProfileFieldType::BOOLEAN:
-                $sql = "".$name." tinyint(1) unsigned NOT NULL default ".bkint($defvalue);
-                break;
-            case UProfileFieldType::INTEGER:
-                $sql = "".$name." int(".$size.") ".$unsignedStr." NOT NULL default ".intval($defvalue)."";
-                break;
-            case UProfileFieldType::STRING:
-                $sql = "".$name." varchar(".$size.") NOT NULL default '".$defvalue."'";
-                break;
-            case UProfileFieldType::DOUBLE:
-                $sql = "".$name." double(".$size.") ".$unsignedStr." NOT NULL default ".doubleval($defvalue)."";
-                break;
-            case UProfileFieldType::TEXT:
-                $sql = "".$name." TEXT";
-                break;
-            case UProfileFieldType::DATETIME:
-                $sql = "".$name." int(10) unsigned NOT NULL default 0";
-                break;
-            case UProfileFieldType::ENUM:
-                $sql = "".$name." int(".$size.") unsigned NOT NULL default 0";
-                break;
-            case UProfileFieldType::TABLE:
-                $sql = "".$name."id int(".$size.") unsigned NOT NULL default 0";
-                break;
-        }
-        if (empty($sql)){
-            return;
-        }
-        $sql = "ALTER TABLE ".$db->prefix."user ADD ".$sql." COMMENT '".$title."'";
-        $db->query_write($sql);
-    }
-
-    public static function FieldInfoAppend(Ab_Database $db, $name, $title, $type, $ops){
-        $sql = "
-			INSERT INTO ".$db->prefix."upfl_field (fieldname, title, fieldtype, fieldaccess, options, ord) VALUES (
-				'".bkstr($name)."',
-				'".bkstr($title)."',
-				".bkint($type).",
-				".bkint($ops['access']).",
-				'".bkstr($ops['options'])."',
-				".bkint($ops['order'])."
-			)
-		";
-        $db->query_write($sql);
-    }
-
-    public static function FieldRemove(Ab_Database $db, $name){
-        $sql = "
-			ALTER TABLE ".$db->prefix."user DROP ".bkstr($name)."
-		";
-        $db->query_write($sql);
-        $sql = "
-			DELETE FROM ".$db->prefix."upfl_field
-			WHERE fieldname='".bkstr($name)."'
-		";
-        $db->query_write($sql);
-    }
-
-    public static function FieldAccessUpdate(Ab_Database $db, $name, $access){
-        $sql = "
-			UPDATE ".$db->prefix."upfl_field
-			SET fieldaccess=".bkint($access)."
-			WHERE fieldname='".bkstr($name)."'
-			LIMIT 1
-		";
-        $db->query_write($sql);
-    }
-
 }