+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### 0.1.7
+
+- add user profile field: github, facebook, vk and etc.
+- update by new URating module
+- show user profile by username
+- bugfix: user profile page
+
 ### 0.1.6
 
 - license changed to MIT